FR-1683: Refactor BSP classes to use common components (#362)
