--- src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/document/Addressee.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.document;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Builder
-@EqualsAndHashCode
-public class Addressee {
-
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("formattedAddress")
-    private String formattedAddress;
-}
