+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/helper/DocumentHelperTest.java
@@ -3,10 +3,10 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.helper;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
+import uk.gov.hmcts.reform.bsp.common.model.document.CtscContactDetails;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.CtscContactDetails;
 
 import java.io.InputStream;
 import java.util.HashMap;
